Remove Python 2-only code paths.
